requirements.txt
scikit-image==0.16.2
scikit-learn==1.0.2
scipy==1.3.1
tqdm
torch==1.12.0
torchvision==0.13.0
numpy==1.21.6
imageio==2.21.2
faiss-gpu==1.7.2
pyyaml==6.0